fixed MSVC compile warnings

+++ kernels/bvh/bvh.h
@@ -1095,13 +1095,13 @@ namespace embree
       assert(!((size_t)node & align_mask));
       ssize_t node_offset = (ssize_t)node-(ssize_t)base;
       assert(node_offset != 0);
-      assert(node_offset >= -0x80000000 && node_offset <= 0x7fffffff);
+      assert(node_offset >= -ssize_t(0x80000000) && node_offset <= 0x7fffffff);
       return (unsigned int)node_offset | tyQuantizedNode;
     }
 
     static __forceinline int encodeQuantizedLeaf(size_t base, size_t node) {
       ssize_t leaf_offset = (ssize_t)node-(ssize_t)base;
-      assert(leaf_offset >= -0x80000000 && leaf_offset <= 0x7fffffff);
+      assert(leaf_offset >= -ssize_t(0x80000000) && leaf_offset <= 0x7fffffff);
       return (int)leaf_offset;
     }
